backend/api-service/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
pandas>=2.2,<3.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<1.0
pymupdf>=1.24,<2.0
python-docx>=1.1,<2.0
gunicorn>=22.0,<24.0
google-auth>=2.29,<3.0
gspread>=6.1,<7.0
requests>=2.31,<3.0
httpx>=0.27,<1.0
